==> 02-data-preprocessing.ipynb <==
# This notebook includes data cleaning and type correction steps.

import pandas as pd

file_path = 'Data/Impact_of_Remote_Work_on_Mental_Health.csv'
df = pd.read_csv(file_path)

null_counts = df.isnull().sum()

null_counts

null_counts / df.shape[0] * 100

df['Mental_Health_Condition'].value_counts()

df['Mental_Health_Condition'] = df['Mental_Health_Condition'].fillna("Unknown")

df['Mental_Health_Condition'].value_counts()

null_counts_2 = df.isnull().sum()

null_counts_2 / df.shape[0] * 100

df['Physical_Activity'].value_counts()

df['Physical_Activity'] = df['Physical_Activity'].fillna("Occasional or Never")

df['Physical_Activity'].value_counts()

null_counts_3 = df.isnull().sum()

null_counts_3 / df.shape[0] * 100

df.info()

unique_values = {column: df[column].unique() for column in df.select_dtypes(include='object').columns}

unique_values

categorical_columns = [
    'Gender', 'Job_Role', 'Industry', 'Work_Location', 'Stress_Level',
    'Mental_Health_Condition', 'Access_to_Mental_Health_Resources',
    'Productivity_Change', 'Satisfaction_with_Remote_Work', 'Physical_Activity',
    'Sleep_Quality', 'Region'
]

for column in categorical_columns:
    df[column] = df[column].astype('category')

df.info()

df.to_csv('Data/Cleaned_Impact_of_Remote_Work_on_Mental_Health.csv', index = False)
