==> 07-is-remote-work-stress-balanced-prediction.ipynb <==
# This notebook includes model developement with a new target variable as Is_RemoteWork_Stress_Balanced

import pandas as pd
from sklearn.model_selection import train_test_split, cross_val_score
from sklearn.metrics import classification_report, confusion_matrix, accuracy_score

file_path = 'Data/Cleaned_Impact_of_Remote_Work_on_Mental_Health.csv'
df = pd.read_csv(file_path)

#### Data cleanup for unrelated values for Satisfaction_with_Remote_Work

df = df.loc[df['Work_Location'] != 'Onsite']

#### Remove outliers

df = df.drop(df[(df['Satisfaction_with_Remote_Work'] == 'Unsatisfied') & (df['Work_Life_Balance_Rating'] > 3)].index)

#### Encoding

df['Stress_Level_Encoded'] = df['Stress_Level'].apply(lambda x: 2 if x == 'Low' else (1 if x == 'Medium' else 0))

def is_remote_work_stress_balanced(row):
    if row['Satisfaction_with_Remote_Work'] == 'Satisfied' and (row['Stress_Level_Encoded'] > 1 or row['Work_Life_Balance_Rating'] > 3):
        return 1
    elif row['Satisfaction_with_Remote_Work'] == 'Neutral' and (row['Stress_Level_Encoded'] > 1 or row['Work_Life_Balance_Rating'] > 3):
        return 1
    else:
        return 0

df['Is_RemoteWork_Stress_Balanced'] = df.apply(is_remote_work_stress_balanced, axis = 1)

features_to_drop = ['Employee_ID', 'Satisfaction_with_Remote_Work', 'Is_RemoteWork_Stress_Balanced', 'Stress_Level', 'Work_Location', 
                    'Social_Isolation_Rating', 'Physical_Activity', 'Mental_Health_Condition', 'Job_Role', 'Industry', 
                    'Access_to_Mental_Health_Resources', 'Productivity_Change', 'Region']

X = pd.get_dummies(df.drop(features_to_drop, axis = 1), drop_first = True)
y = df['Is_RemoteWork_Stress_Balanced']

from sklearn.model_selection import train_test_split

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size = 0.2, random_state = 42)

class_distribution = df['Is_RemoteWork_Stress_Balanced'].value_counts()
print(class_distribution)

## Using RandomForest

#### Tuning

from bayes_opt import BayesianOptimization
from sklearn.model_selection import cross_val_score
from sklearn.model_selection import StratifiedKFold
from sklearn.ensemble import RandomForestClassifier
import numpy as np

def RandomForest_Bayesian(n_estimators, max_depth, min_samples_split, min_samples_leaf):
    model = RandomForestClassifier(
        max_depth = int(max_depth),
        n_estimators = int(n_estimators),
        min_samples_split = int(min_samples_split),
        min_samples_leaf = int(min_samples_leaf),
        class_weight = 'balanced',
        random_state = 42
    )
    
    # Performing cross-validation
    cv = StratifiedKFold(n_splits = 5, shuffle = True, random_state = 42)
    scores = cross_val_score(model, X_train, y_train, cv = cv, scoring = 'f1')
    return np.mean(scores)

param_bounds = {
    'n_estimators': (50, 300),
    'max_depth': (3, 20),
    'min_samples_split': (2, 10),
    'min_samples_leaf': (1, 5)
}

optimizer = BayesianOptimization(
    f = RandomForest_Bayesian,
    pbounds = param_bounds,
    random_state = 42,
    verbose = 2
)

optimizer.maximize(init_points = 20, n_iter = 50)
print("Best Parameters:", optimizer.max)

#### Train

rf_model = RandomForestClassifier(
    n_estimators = int(optimizer.max['params']['n_estimators']),
    max_depth = int(optimizer.max['params']['max_depth']),
    min_samples_split = int(optimizer.max['params']['min_samples_split']),
    min_samples_leaf = int(optimizer.max['params']['min_samples_leaf']),
    class_weight = 'balanced',
    random_state = 42
)

rf_model.fit(X_train, y_train)

#### Predict

y_pred = rf_model.predict(X_test)

print("\nClassification Report:\n", classification_report(y_test, y_pred))
print("Confusion Matrix:\n", confusion_matrix(y_test, y_pred))

#### Feature Importance

feature_importance = pd.DataFrame({
    'Feature': X.columns,
    'Importance': rf_model.feature_importances_
}).sort_values(by = 'Importance', ascending = False)

print(feature_importance)
